+++ src/integrationTest/java/fll/web/FullTournamentTest.java
@@ -645,7 +645,7 @@ public class FullTournamentTest {
         + "_teams.csv");
     // write the teams out to a file
     try (Writer writer = Files.newBufferedWriter(teamsFile, Utilities.DEFAULT_CHARSET)) {
-      try (CSVWriter csvWriter = new CSVWriter(writer)) {
+      try (CSVWriter csvWriter = Utilities.createCSVWriter(writer)) {
         csvWriter.writeNext(new String[] { "team_name", "team_number", "affiliation", "award_group", "judging_group",
                                            "tournament" });
         final Map<Integer, TournamentTeam> sourceTeams = Queries.getTournamentTeams(testDataConnection,